Se implementan los métodos para realizar la sustitución de resultados de muestreo.

+++ WebAPI/Controllers/v1/Operacion/Limites.cs
@@ -1,4 +1,5 @@
 ﻿using Application.DTOs;
+using Application.Enums;
 using Application.Features.Operacion.SustitucionLimites.Commands;
 using Domain.Settings;
 using Microsoft.AspNetCore.Mvc;
@@ -12,22 +13,25 @@ namespace WebAPI.Controllers.v1.Operacion
         [DisableRequestSizeLimit]
         public async Task<IActionResult> Post([FromForm] ParametrosSustitucionLimitesDto parametrosSustitucionLimites)
         {
-            var sustítucionCorrecta = await Mediator.Send(new SustitucionMaximoComunCommand { ParametrosSustitucion = parametrosSustitucionLimites });
-
-            string filePath = string.Empty;
-
-            if (parametrosSustitucionLimites.Archivo?.Length > 0)
+            if (parametrosSustitucionLimites.OrigenLimites == (int)TipoSustitucionLimites.TablaTemporal)
             {
-                filePath = Path.GetTempFileName();
-                using var stream = System.IO.File.Create(filePath);
-                await parametrosSustitucionLimites.Archivo.CopyToAsync(stream);
+                string filePath = string.Empty;
+
+                if (parametrosSustitucionLimites.Archivo?.Length > 0)
+                {
+                    filePath = Path.GetTempFileName();
+                    using var stream = System.IO.File.Create(filePath);
+                    await parametrosSustitucionLimites.Archivo.CopyToAsync(stream);
+                }
+
+                FileInfo fileInfo = new(filePath);
+                ExcelService.Mappings = ExcelLimitesComunes.keyValues;
+                var registros = ExcelService.Import<LimiteMaximoComunDto>(fileInfo, "Límites 2012-2022");
+                System.IO.File.Delete(filePath);
+                parametrosSustitucionLimites.LimitesComunes = registros;
             }
 
-            FileInfo fileInfo = new(filePath);
-            ExcelService.Mappings = ExcelLimitesComunes.keyValues;
-            var registros = ExcelService.Import<LimiteMaximoComunDto>(fileInfo, "Límites 2012-2022");
-            System.IO.File.Delete(filePath);
-
+            await Mediator.Send(new SustitucionMaximoComunCommand { ParametrosSustitucion = parametrosSustitucionLimites });
 
             return Ok();
         }
